requirements.txt
click==6.7
Flask==0.12.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pymongo==3.5.1
python-dotenv==0.7.1
Werkzeug==0.12.2
